+++ packages/mui-components/src/components/atoms/Button/types.ts
@@ -1,0 +1,8 @@
+export type ButtonType = {
+  color: ButtonColour;
+  variant: ButtonVariant;
+};
+
+export type ButtonColour = 'primary' | 'secondary' | 'warning' | 'inherit' | 'success' | 'error' | 'info' | undefined;
+
+export type ButtonVariant = 'contained' | 'outlined' | 'text' | undefined;