+++ packages/mui-components/src/components/atoms/index.ts
@@ -1,1 +1,4 @@
 export * from './FormErrorText';
+export * from './Button';
+export * from './SROnly';
+export * from './Loader';