+++ packages/mui-components/src/index.ts
@@ -1,3 +1,6 @@
+/// <reference types="vite-plugin-svgr/client" />
+
 export * from './components';
+export * from './theme';
 export * from './hooks';
 export * from './test-utils';