+++ packages/mui-components/src/theme/components/MuiBreadcrumbs/index.ts
@@ -5,7 +5,7 @@ export const MuiBreadcrumbs: Components<Theme>['MuiBreadcrumbs'] = {
     root: ({ theme }) => ({
       fontSize: '0.875rem',
       lineHeight: 1.4,
-      color: theme.palette.secondary.main,
+      color: theme.palette.secondary.dark,
     }),
 
     separator: {