+++ packages/mui-components/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @batoanng/mui-components
 
+## 3.0.3
+
+### Patch Changes
+
+- Add Button, Loader, SROnly
+
 ## 3.0.2
 
 ### Patch Changes